requirements.txt
requests~=2.31.0
python-slugify~=8.0.1
selenium~=4.11.2